+++ src/TypeScript/TypeScript.fs
@@ -34,64 +34,17 @@ type TsType =
 and Variable = 
   { name : string
     varType : TsType option
-    Comment: Comment option
+    comment: string list
     value : Value option
     declare: bool
     optional: bool }
   static member Create(name, ?varType, ?comment, ?value, ?declare, ?optional) = 
     { Variable.name = name
       varType = varType
-      Comment = comment 
+      comment = defaultArg comment []
       value = value
       declare = defaultArg declare false
       optional = defaultArg optional false }
-
-and Comment =
-  { displayName: string
-    setName: string
-    label: string
-    isPrimaryId: bool
-    colType: XrmAttributeType option
-    targetEntitySets: (string * string * string)[] option
-    relType: RelType option
-    tab: string
-    intersectTable: string
-    link: string }
-  static member Create(?displayName, ?setName, ?intersectTable, ?label, ?isPrimaryId, ?colType, ?tes, ?relType, ?tab, ?link) =
-    { displayName = defaultArg displayName ""
-      setName = defaultArg setName ""
-      label = defaultArg label ""
-      isPrimaryId = defaultArg isPrimaryId false
-      intersectTable = defaultArg intersectTable ""
-      colType = colType
-      targetEntitySets = tes
-      relType = relType
-      tab = defaultArg tab ""
-      link = defaultArg link "" }
-  member c.ToCommentStrings() =
-    let dsLine = if String.IsNullOrWhiteSpace c.displayName then None else Some $"**{c.displayName.Trim()}**"
-    let setNameLine = if String.IsNullOrWhiteSpace c.setName then None else Some $"Set Name: `{c.setName.Trim()}`"
-    let tabLine = if String.IsNullOrWhiteSpace c.tab then None else Some $"Tab: {c.tab.Trim()}"
-    let labelLine = if String.IsNullOrWhiteSpace c.label then None else Some $"Label: {c.label.Trim()}"
-    let isPrimaryIdLine = if c.isPrimaryId then Some $"Primary ID" else None
-    let colTypeLine = c.colType |> Option.map (fun t -> $"Column Type: {t}")
-    let intersectTableLine = if String.IsNullOrWhiteSpace c.intersectTable then None else Some $"Intersect Table: `{c.intersectTable.Trim()}`"
-    let tableLine =
-      match c.targetEntitySets with
-      | None | Some [||] -> None
-      | Some tes ->
-        let maxDisplay = 5
-        let shown = tes |> Array.truncate maxDisplay |> Array.map (fun (ln, _, dn) -> $"{dn} (`{ln}`)") |> String.concat " | "
-        let formatted = if tes.Length <= maxDisplay then shown else $"{shown} | +{tes.Length - maxDisplay} more"
-        Some $"Table: {formatted}"
-    let relTypeLine = c.relType |> Option.map (fun t -> $"Relationship Type: {t}")
-    let linkLine = if String.IsNullOrWhiteSpace c.link then None else Some (sprintf "{@link %s}" (c.link.Trim()))
-
-    let lines = [ dsLine; setNameLine; tabLine; labelLine; isPrimaryIdLine; colTypeLine; tableLine; relTypeLine; linkLine; intersectTableLine ] |> List.choose id
-    match lines with
-    | [] -> []
-    | [ line ] -> [ $"/** {line} */" ]
-    | _ -> [ "/**" ] @ (lines |> List.map (sprintf " * %s  ")) @ [ " */" ]
 
 type ExportType = 
   | Regular
@@ -100,27 +53,27 @@ type ExportType =
 type TsEnum = 
   { name : string
     vals : (string * int option) list
-    comment: Comment option
+    comment: string list
     declare : bool
     constant: bool
     export : bool }
   static member Create(name, ?vals, ?comment, ?constant, ?declare, ?export) = 
     { TsEnum.name = name
       vals = defaultArg vals []
-      comment = comment
+      comment = defaultArg comment []
       declare = defaultArg declare false
       constant = defaultArg constant true
       export = defaultArg export false }
 
 type Function = 
   { name : string
-    comment : Comment option
+    comment : string list
     args : Variable list
     returnType : TsType option
     expr : string list }
   static member Create(name, ?args, ?returnType, ?comment, ?expr) = 
     { Function.name = name
-      comment = comment
+      comment = defaultArg comment []
       args = defaultArg args []
       returnType = returnType
       expr = defaultArg expr [] }
@@ -145,14 +98,14 @@ type Class =
 
 type Interface = 
   { name : string
-    comment : Comment option
+    comment : string list
     extends : string list
     export : ExportType
     vars : Variable list
     funcs : Function list }
   static member Create(name, ?comment, ?extends, ?export, ?vars, ?funcs) = 
     { Interface.name = name
-      comment = comment
+      comment = defaultArg comment []
       extends = defaultArg extends []
       export = defaultArg export Regular
       vars = defaultArg vars []